core: bundle all native modules
They are part of libsjs now (the native parts of the builtin modules, that is).
Instead of using require, they are accessed via `_system.binding()`. I chose
not to document this since it's internal and only to be used by the stdlib.

Addons will continue to be shared libraries (.jsdll files) and using require.